backend/requirements.txt
# Web Framework
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6

# Database
sqlalchemy==2.0.23
alembic==1.12.1
psycopg2-binary==2.9.9

# Validation
pydantic==2.5.0
pydantic-settings==2.1.0

# HTTP Client
httpx==0.25.1

# Authentication
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4

# Caching/Redis
redis==5.0.1

# Reliability
tenacity==8.2.3

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1

# CORS
python-cors==1.0.0

# Utilities
python-dotenv==1.0.0

# AI Provider SDKs
openai==1.54.0
google-generativeai==0.8.3
